backend/image_generator/requirements.txt
Flask==2.3.0
torch>=2.0.0
diffusers>=0.18.0
Pillow>=9.0.0
transformers>=4.0.0
accelerate>=0.12.0
bitsandbytes>=0.39.0